requirements.txt
python-dotenv
Flask
langchain_openai
langchain
openai